Add compatibility with casino system
