+++ src/main/java/ch/bbbaden/casinopalace/roulette/game/Datenbank.java
@@ -5,18 +5,38 @@
  */
 package ch.bbbaden.casinopalace.roulette.game;
 
+import ch.bbbaden.casinopalace.common.Casino;
+import ch.bbbaden.casinopalace.common.exception.UserDoesNotExistException;
+
+import java.io.IOException;
+import java.math.BigDecimal;
+import java.math.BigInteger;
+import java.math.RoundingMode;
+
 /**
  *
  * @author gabri
  */
 public class Datenbank {
-    private int konto = 5000;
+    private final Casino casino;
+
+    public Datenbank(Casino casino) {
+        this.casino = casino;
+    }
 
     public int getKonto() {
-        return konto;
+        return casino.getCurrentUser().getChips().setScale(0, RoundingMode.DOWN).intValue();
     }
 
     public void setKonto(int konto) {
-        this.konto = konto;
+        BigDecimal oldValue = casino.getCurrentUser().getChips();
+        BigInteger integer = oldValue.toBigInteger();
+        BigDecimal fraction = oldValue.subtract(new BigDecimal(integer));
+        casino.getCurrentUser().setChips(new BigDecimal(konto).add(fraction));
+        try {
+            casino.updateUser(casino.getCurrentUser());
+        } catch (IOException | UserDoesNotExistException e) {
+            e.printStackTrace();
+        }
     }
 }
