+++ src/main/java/ch/bbbaden/casinopalace/roulette/game/ProgramInfoController.java
@@ -5,17 +5,12 @@
  */
 package ch.bbbaden.casinopalace.roulette.game;
 
-import ch.bbbaden.casinopalace.roulette.menu.Roulette_MenuController;
 import java.net.URL;
 import java.util.ResourceBundle;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.Spinner;
-import javafx.scene.control.SpinnerValueFactory;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
