--- src/main/java/ch/bbbaden/casinopalace/roulette/menu/Roulette_Menu.fxml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.Pane?>
-<?import javafx.scene.layout.RowConstraints?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="450.0" styleClass="mainFxmlClass" stylesheets="@roulette_menu.css" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ch.bbbaden.casinopalace.roulette.menu.Roulette_MenuController">
-   <children>
-      <Pane layoutX="96.0" layoutY="99.0" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="-Infinity" minWidth="-Infinity" prefHeight="247.0" prefWidth="258.0">
-         <children>
-            <ImageView fx:id="rouletteimage" fitHeight="260.0" fitWidth="258.0" layoutY="-12.0" pickOnBounds="true" preserveRatio="true" />
-         </children>
-      </Pane>
-      <Button fx:id="button" defaultButton="true" layoutX="155.0" layoutY="346.0" mnemonicParsing="false" onAction="#clickStart" prefHeight="40.0" prefWidth="140.0" text="Start" textFill="YELLOW">
-         <font>
-            <Font name="Trebuchet MS Bold" size="35.0" />
-         </font>
-      </Button>
-      <GridPane layoutX="96.0" layoutY="14.0" prefHeight="66.0" prefWidth="258.0">
-        <columnConstraints>
-          <ColumnConstraints hgrow="SOMETIMES" maxWidth="1.7976931348623157E308" minWidth="10.0" prefWidth="100.0" />
-        </columnConstraints>
-        <rowConstraints>
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-        </rowConstraints>
-         <children>
-            <ImageView fx:id="titleimage" fitHeight="121.0" fitWidth="281.0" pickOnBounds="true" preserveRatio="true">
-               <image>
-                  <Image url="@../../../../../../resources/images/roulette.png" />
-               </image>
-            </ImageView>
-         </children>
-      </GridPane>
-      <Label layoutX="14.0" layoutY="14.0" text="Kontostand:" />
-      <Label fx:id="kontostand" layoutX="82.0" layoutY="14.0" />
-   </children>
-</AnchorPane>
